requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic[email]
httpx
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Vector Database
pgvector==0.2.4

# AI/ML
sentence-transformers==2.3.1
instructor==0.4.5
ollama==0.1.6
openai==1.10.0

# Background Tasks
apscheduler==3.10.4
celery==5.3.6
redis==5.0.1

# API Integrations
httpx==0.25.2
requests==2.31.0



# LinkedIn API
linkedin-api==2.2.0

# YouTube API
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Email
sendgrid==6.11.0
python-dotenv==1.0.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# DNS Validation
dnspython==2.5.0

# Storage
boto3==1.34.34
python-multipart==0.0.6

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pillow==10.2.0
numpy==1.26.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0